importruns: extract import logic to separate concerns
Move the run import functionality and associated event logging out of
the management command into a dedicated module. This decouples the
business logic from the CLI, improves separation of concerns, and
enables reuse.
